Update to Python3 and upgrade simulator to use multiprocessing API
- Update python code to Python3
- Update setup.py file
- Upgrade the simulator

+++ openvisualizer/utils.py
@@ -85,7 +85,7 @@ def format_buf(buf):
 
 def format_ipv6_addr(addr):
     # group by 2 bytes
-    addr = [buf2int(addr[2 * i:2 * i + 2]) for i in range(len(addr) / 2)]
+    addr = [buf2int(addr[2 * i:2 * i + 2]) for i in range(int(len(addr) / 2))]
     return ':'.join(["%x" % b for b in addr])
 
 
@@ -117,7 +117,7 @@ def hex2buf(s):
 
     return_val = []
 
-    for i in range(len(s) / 2):
+    for i in range(int(len(s) / 2)):
         real_idx = i * 2
         return_val.append(int(s[real_idx:real_idx + 2], 16))
